Add Backspace Button & Decimal Button Template
Made a backspace button that deletes last digit & a decimal button that adds "," to the screen,
but it doesn't work when it comes to end math for some reason.

I also added a list of logic things that I need to fix still. It's located in the end of calculating.js file.